+++ backend/app/views.py
@@ -53,6 +53,17 @@ def gestures_to_sound(request):
 
     return render(request, 'index.html', context)
 
+def text_shape_analysis(request):
+    """ Text Shape Analysis Page """
+
+    context = {
+        'page_metadata': {
+            'title': 'Shape Analysis'
+        },
+        'component_name': 'TextShapeAnalysis'
+    }
+
+    return render(request, 'index.html', context)
 
 ################################################################################
 # Boilerplate - just for examples
